Crear la tool autogit (#64)
This pull request introduces a new script `autogit.sh` to automate Git
operations in a non-UI environment. The script is designed to facilitate
continuous integration by committing, pulling, and pushing changes in a
loop.

Key changes:

*
[`tools/bash/autogit.sh`](diffhunk://#diff-f638675c8e0e3192cf253318bd7fa5e28186add28a4fc4deb7f6047fb1ecf732R1-R28):
Added a new Bash script to automate Git operations. The script commits
all local changes with a generic message, pulls the latest changes while
resolving conflicts by keeping the remote version, and pushes the
changes to the remote repository. It runs in an infinite loop with a
5-minute sleep interval between executions.

Fixes #61