Add work done logging to statistics service (#91)
This pull request includes several changes to enhance the functionality
and data tracking of the statistics service. The most important changes
include adding a new parameter to the `doWork` method, introducing a
method to retrieve the work done log, and updating related classes to
utilize these new functionalities.

### Enhancements to `IStatisticsService`:

* Added a `task` parameter to the `doWork` method to track which task
the work units are associated with.
* Introduced a new abstract method `getWorkDoneLog` to retrieve the log
of work done.

### Updates in `StatisticsService`:

* Modified the `doWork` method to include logging of work done with
timestamps and task descriptions.
* Implemented the `getWorkDoneLog` method to return the work done log.

### Changes in related service classes:

* Updated `workCommand` in `TelegramReportingService` to pass the `task`
parameter when calling `doWork`.
* Enhanced `get_list_stats` in `TelegramTaskListManager` to include the
work done log in the statistics message.

### Utility updates:

* Added a `from_int` method to `TimePoint` class to convert timestamps
back to `TimePoint` objects.

Closes #15